+++ example/gap_dedicated_bonding.c
@@ -77,6 +77,13 @@ int btstack_main(int argc, const char * argv[]){
     hci_event_callback_registration.callback = &packet_handler;
     hci_add_event_handler(&hci_event_callback_registration);
 
+    l2cap_init();
+
+#ifdef ENABLE_BLE
+    // Initialize LE Security Manager. Needed for cross-transport key derivation
+    sm_init();
+#endif
+
     // turn on!
     hci_power_control(HCI_POWER_ON);